Add back same drive Directory.Move

+++ FreeMove/IO/MoveOperation.cs
@@ -11,6 +11,7 @@ namespace FreeMove.IO
     class MoveOperation : IOOperation
     {
         string pathFrom;
+        string pathTo;
         bool sameDrive; //TODO: use System.IO.Move if the files are on the same drive
         CancellationTokenSource cts = new CancellationTokenSource();
         CopyOperation innerCopy;
@@ -18,6 +19,7 @@ namespace FreeMove.IO
         public MoveOperation(string pathFrom, string pathTo)
         {
             this.pathFrom = pathFrom;
+            this.pathTo = pathTo;
             sameDrive = string.Equals(Path.GetPathRoot(pathFrom), Path.GetPathRoot(pathTo), StringComparison.OrdinalIgnoreCase);
 
             innerCopy = new CopyOperation(pathFrom, pathTo);
@@ -35,29 +37,47 @@ namespace FreeMove.IO
             innerCopy.Start += (sender, e) => OnStart(e);
             try
             {
-                try
+                if (sameDrive)
                 {
-                    await innerCopy.Run();
+                    try
+                    {
+                        await Task.Run(() => Directory.Move(pathFrom, pathTo), cts.Token);
+                    }
+                    catch (Exception e) when (!(e is OperationCanceledException))
+                    {
+                        throw new MoveFailedException("Exception encountered while moving on the same drive", e);
+                    }
                 }
-                catch (Exception e) when (!(e is OperationCanceledException)) // Wrap inner exceptions to signal which phase failed
+                else
                 {
-                    throw new CopyFailedException("Exception encountered while copying directory", e);
-                }
+                    try
+                    {
+                        await innerCopy.Run();
+                    }
+                    catch (Exception e) when (!(e is OperationCanceledException)) // Wrap inner exceptions to signal which phase failed
+                    {
+                        throw new CopyFailedException("Exception encountered while copying directory", e);
+                    }
 
-                cts.Token.ThrowIfCancellationRequested();
-                try
-                {
-                    Directory.Delete(pathFrom, true);
-                }
-                catch (Exception e)
-                {
-                    throw new DeleteFailedException("Exception encountered while removing duplicate files in the old location", e);
+                    cts.Token.ThrowIfCancellationRequested();
+                    try
+                    {
+                        Directory.Delete(pathFrom, true);
+                    }
+                    catch (Exception e)
+                    {
+                        throw new DeleteFailedException("Exception encountered while removing duplicate files in the old location", e);
+                    }
                 }
             }
             finally
             {
                 OnEnd(new EventArgs());
             }
+        }
+        public class MoveFailedException : Exception
+        {
+            public MoveFailedException(string message, Exception innerException) : base(message, innerException) { }
         }
         public class CopyFailedException : Exception
         {
